requirements.txt
django>=4.2,<5.0
djangorestframework
django-cors-headers
django-allauth
django-auditlog
django-celery-results
